requirements.txt
asgiref==3.2.7
Babel==2.8.0
Django==3.0.6
django-autoslug==1.9.7
django-autoslug-iplweb==1.9.4.dev0
django-cities-light==3.6.0
django-countries==6.1.2
django-debug-toolbar==2.2
django-leaflet==0.26.0
django-model-utils==4.0.0
django-phonenumber-field==4.0.0
django-querysetsequence==0.12
Faker==4.1.1
geographiclib==1.50
geopy==1.22.0
lxml==4.5.1
phonenumbers==8.12.5
Pillow==7.1.2
progressbar2==3.51.3
psycopg2==2.8.5
python-dateutil==2.8.1
python-gettext==4.0
python-memcached==1.59
python-stdnum==1.13
python-utils==2.4.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
tqdm==4.46.1
Unidecode==1.1.1
